+++ packages/react-hooks/src/ctx/PayWithAsset.tsx
@@ -5,10 +5,9 @@ import type { BN } from '@polkadot/util';
 import type { AssetInfoComplete } from '../types.js';
 import type { PayWithAsset } from './types.js';
 
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { useApi, useAssetIds, useAssetInfos } from '@polkadot/react-hooks';
-import { formatNumber } from '@polkadot/util';
 
 interface Props {
   children?: React.ReactNode;
@@ -45,7 +44,7 @@ function PayWithAssetProvider ({ children }: Props): React.ReactElement<Props> {
     () => (assetInfos
       ?.filter((i): i is AssetInfoComplete =>
         !!(i.details && i.metadata) && !i.details.supply.isZero() && !!i.details?.toJSON().isSufficient)
-      .sort((a, b) => a.id.cmp(b.id))) || [],
+    ) || [],
     [assetInfos]
   );
 
@@ -53,7 +52,7 @@ function PayWithAssetProvider ({ children }: Props): React.ReactElement<Props> {
     () => [
       { text: `${nativeAsset} (Native)`, value: nativeAsset },
       ...completeAssetInfos.map(({ id, metadata }) => ({
-        text: `${metadata.name.toUtf8()} (${formatNumber(id)})`,
+        text: `${metadata.name.toUtf8()} (${id.toString()})`,
         value: id.toString()
       }))],
     [completeAssetInfos, nativeAsset]
@@ -70,9 +69,15 @@ function PayWithAssetProvider ({ children }: Props): React.ReactElement<Props> {
     api.registry.metadata.extrinsic.signedExtensions.some(
       (a) => a.identifier.toString() === 'ChargeAssetTxPayment'
     ) &&
-    api.tx.assetConversion && completeAssetInfos.length > 0,
-  [api.registry.metadata.extrinsic.signedExtensions, api.tx.assetConversion, completeAssetInfos.length]
+    !!api.tx.assetConversion &&
+    !!api.call.assetConversionApi &&
+    completeAssetInfos.length > 0,
+  [api.call.assetConversionApi, api.registry.metadata.extrinsic.signedExtensions, api.tx.assetConversion, completeAssetInfos.length]
   );
+
+  useEffect(() => {
+    return () => setSelectedFeeAsset(null);
+  });
 
   const values: PayWithAsset = useMemo(() => {
     return {