+++ packages/react-signer/src/PayWithAsset.tsx
@@ -1,22 +1,28 @@
-// Copyright 2017-2025 @polkadot/react-components authors & contributors
+// Copyright 2017-2025 @polkadot/react-signer authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import type { DropdownItemProps } from 'semantic-ui-react';
+import type { ExtendedSignerOptions } from './types.js';
 
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { Dropdown } from '@polkadot/react-components';
+import { Dropdown, Modal } from '@polkadot/react-components';
 import { useApi, usePayWithAsset } from '@polkadot/react-hooks';
+import { getFeeAssetLocation } from '@polkadot/react-hooks/utils/getFeeAssetLocation';
 import { BN } from '@polkadot/util';
 
 import { useTranslation } from './translate.js';
 
-const PayWithAsset = () => {
+interface Props {
+  onChangeFeeAsset: React.Dispatch<React.SetStateAction<ExtendedSignerOptions>>
+}
+
+const PayWithAsset = ({ onChangeFeeAsset }: Props) => {
   const { t } = useTranslation();
   const { api } = useApi();
-  const [selectedAssetValue, setSelectedAssetValue] = useState('0');
+  const [selectedAssetValue, setSelectedAssetValue] = useState('-1');
 
-  const { assetOptions, isDisabled, onChange } = usePayWithAsset();
+  const { assetOptions, isDisabled, onChange, selectedFeeAsset } = usePayWithAsset();
 
   const nativeAsset = useMemo(
     () => api.registry.chainTokens[0],
@@ -46,15 +52,29 @@ const PayWithAsset = () => {
     }
   }, [assetOptions, selectedAssetValue, nativeAsset]);
 
+  useEffect(() => {
+    if (selectedFeeAsset) {
+      onChangeFeeAsset((e) =>
+        ({
+          ...e,
+          assetId: getFeeAssetLocation(api, selectedFeeAsset),
+          feeAsset: selectedFeeAsset
+        })
+      );
+    }
+  }, [api, onChangeFeeAsset, selectedFeeAsset]);
+
   return (
-    <Dropdown
-      isDisabled={isDisabled}
-      label={t('asset to pay the fee')}
-      onChange={onSelect}
-      onSearch={onSearch}
-      options={assetOptions}
-      value={selectedAssetValue}
-    />
+    <Modal.Columns hint={t('By selecting this option, the transaction fee will be automatically deducted from the specified asset, ensuring a seamless and efficient payment process.')}>
+      <Dropdown
+        isDisabled={isDisabled}
+        label={t('asset to pay the fee')}
+        onChange={onSelect}
+        onSearch={onSearch}
+        options={assetOptions}
+        value={selectedAssetValue}
+      />
+    </Modal.Columns>
   );
 };